Check latest release on GitHub and pupup message for updates
Every time the launcher starts, it will now check GitHub for the latest
release and compare the tag to the `VERSION` macro, which must be set
before compiling. A popup will notify the user with a link to the latest
GitHub release page. This behavior can be disabled with a checkbox in
the options menu.